boutique_ai_saas/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
Pillow>=10.0
requests>=2.31
python-dotenv>=1.0
qrcode>=7.4
